room_booking_system/requirements.txt
alembic==1.14.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.6.0
django-mysql==4.15.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
Flask==2.2.3
Flask-Cors==3.0.9
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.5.1
Mako==1.3.8
MarkupSafe==2.1.1
msgpack==1.1.0
mysqlclient==2.1.1
pika==1.3.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.28.1
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
SQLAlchemy==1.4.46
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
vine==5.1.0
wait-for-it==2.3.0
wcwidth==0.2.13
Werkzeug==2.2.3
whitenoise==6.9.0
zope.interface==7.2